+++ NEWS.md
@@ -2,6 +2,7 @@
 # QBMS 0.9.0 _2022-11-??_
   * Add TerraClimate support to query and retrieve climate data (including the [19 BIOCLIM variables](https://www.worldclim.org/data/bioclim.html)).
   * Add a new gigwa_get_metadata() function to retrieve metadata from GIGWA database.
+  * The get_pedigree_table() function can detect and handle backcross cases properly in the pedigree strings.
   * Implement caching techniques wherever applicable for faster response times, which can improve user experience. Thanks to [Johan Steven Aparicio](https://github.com/AparicioJohan) for initiating this.
   * Improve the performance of multi-page API calls by [optionally enabling asynchronous calls](https://github.com/gaborcsardi/async) to prevent blocking behavior by fetching all requested pages simultaneously. This helps to reduce the waiting time for the user.
   * Fix the error message when calling the get_pedigree_table() function if there is no case of similar genotype names. Thanks to [Johan Steven Aparicio](https://github.com/AparicioJohan) for reporting this.